requirements.txt
psutil==5.9.6
requests==2.31.0
Flask==3.0.0
Flask-PyMongo==2.3.0
python-dotenv==1.0.0